consolidate DND_TYPES constant
