+++ protocol-designer/src/components/DeckSetup/LabwareOverlays/EditLabware.js
@@ -7,7 +7,7 @@ import { getLabwareDisplayName } from '@opentrons/shared-data'
 import { DragSource, DropTarget } from 'react-dnd'
 import { i18n } from '../../../localization'
 import { NameThisLabware } from './NameThisLabware'
-import { DND_TYPES } from './constants'
+import { DND_TYPES } from '../../../constants'
 import {
   openIngredientSelector,
   deleteContainer,
