+++ protocol-designer/src/components/DeckSetup/LabwareOverlays/__tests__/SlotControls.test.js
@@ -4,6 +4,7 @@ import React from 'react'
 import { shallow } from 'enzyme'
 import fixture_96_plate from '@opentrons/shared-data/labware/fixtures/2/fixture_96_plate.json'
 import { MAGNETIC_MODULE_TYPE } from '@opentrons/shared-data'
+import { DND_TYPES } from '../../../../constants'
 import * as labwareModuleCompatibility from '../../../../utils/labwareModuleCompatibility'
 import { START_TERMINAL_ITEM_ID } from '../../../../steplist'
 import { SlotControlsComponent } from '../SlotControls'
@@ -42,6 +43,7 @@ describe('SlotControlsComponent', () => {
       draggedItem: {
         labwareOnDeck,
       },
+      itemType: DND_TYPES.LABWARE,
       customLabwareDefs: {},
     }
