--- protocol-designer/src/components/DeckSetup/LabwareOverlays/constants.js
@@ -1,5 +1,0 @@
-// @flow
-// TODO: BC 2019-01-03 consolidate with DraggableStepItems DND_TYPES
-export const DND_TYPES: { LABWARE: 'LABWARE' } = {
-  LABWARE: 'LABWARE',
-}
